+++ src/routes/v1/index.ts
@@ -1,11 +1,11 @@
 import express, { Router } from 'express';
 import auth from './middlewares/Auth/auth';
-import finalOption from './finalOption';
-import meetInfo from './meetInfo';
+import finalChoice from './finalChoice';
+import invite from './invite';
 
 const router: Router = express.Router();
 router.use('/', auth);
-router.use('/finalOption', finalOption);
-router.use('/meetInfo', meetInfo);
+router.use('/finalChoice', finalChoice);
+router.use('/invite', invite);
 
 export default router;
